[HOT FIX] Add debug libs for osx
Added debug libs for osx currently release libs build on osx do not work
Issue building v8 on osx with libc++ works but building v8dotnetproxy with
libc++ does not work for mono, leads to an bad_access error after executing
an pinvoke call.
Building v8 with stdlibc++ leads to an segmentation fault when excecuting in release mode.
Hot Fix is to add the debug libs which are significant slower!!!